requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.14.0
pillow
numpy

